Consolidate Flamingo on top of Substance
Tracked by https://github.com/kirill-grouchnikov/radiance/issues/10 - move all the relevant code from substance-flamingo to flamingo and remove the code that was emulating the non-Substance look of controls.

Also start fixing some of the demo apps in flamingo-demo to use Substance, and switch kormorant-demo to that + proper Swing threading.

In addition, remove some of the code that was duplicated between Flamingo and Substance.