Introducing Neon
Tracked by https://github.com/kirill-grouchnikov/radiance/issues/11 - consolidate everything related to scalable content in one place so that Substance and Flamingo use a unified approach to working with that.

Switch the relevant Substance APIs to the Neon world (backwards incompatible change that involves renaming some of the classes and moving them into a new package).