Fix StateTransitionMultiTracker concurrency
The calls to trackerMap.remove() in the StateTransitionListener methods in the StateTransitionMultiTracker class escape the synchronization that is applied to other uses of the trackerMap field. A synchronized removeTracker method analogous to the addTracker method fixes this.

If the listener methods are called on a different thread than the other (synchronized) methods, the trackerMap can become corrupted, which can lead to deadlocks.