Bump to 2.0-SNAPSHOT + remove deprecated Flamingo APIs
As part of https://github.com/kirill-grouchnikov/radiance/issues/73

First drop of command projection functionality. A lot of things and demos are probably broken and will be fixed in the next pass. Remove all deprecated APIs so that the evolution can move forward unimpeded.

Also bump Kotlin to 1.3.10 and coroutines to 1.0.1