Trident API changes
Address issues raised in https://github.com/kirill-grouchnikov/radiance/issues/143

Prior to this change, APIs to create and configure Trident timelines did not provide any build-time checks on the designed usage. In addition, some of the APIs also could be used to either reconfigure or reuse a timeline in a way that can lead to poorly specified behavior.

With this change, timelines are created with builders - as in Flamingo's new world. After a timeline is created, it can be play()'d in the same ways as before. It can be queried. But it does not allow changing duration, adding more properties, adding callbacks etc. This provides much stronger guarantees about the expected usage of the APIs.