Creating a JRibbonFrame wihout an application menu causes an NPE
I'm not sure if this is related to #78 or not, but if you create an instance of `JRibbonFrame` and don't set an app menu then this triggers an NPE in the latest 2.0.1 release of Flamingo as well as 2.5-SNAPSHOT on line 471

https://github.com/kirill-grouchnikov/radiance/blob/1c9325c62873c32444d493d667d45800d9621d1c/flamingo/src/main/java/org/pushingpixels/flamingo/internal/ui/ribbon/BasicRibbonUI.java#L471

The problem seems to be that the `applicationMenuButton` is only initialized if it is going to be shown (which makes sense)

https://github.com/kirill-grouchnikov/radiance/blob/1c9325c62873c32444d493d667d45800d9621d1c/flamingo/src/main/java/org/pushingpixels/flamingo/internal/ui/ribbon/BasicRibbonUI.java#L141

This PR fixes two places in `JRibbonFrame` where the `applicationMenuButton` was being accessed even though it was null. Testing seems to suggest that this fully fixes the NPE (or at least I can now create a `JRibbonFrame` without an application menu).