Refactor the internals of command buttons
The newly added `BaseCommandButtonProjection.getPopupMenuPanelProjection()` is used to get the projection for the popup menu panel content. This also allows removing the need to have the `JCommandButton` to know the "difference" between a regular command button and a ribbon application menu button - as well as removing the need for a separate `JColorSelectorCommandButton` extension of `JCommandButton` just to configure its popup content.

For #447