Customizable popup anchor bounds provider in command buttons
This, together with `setShowPopupIcon(false)` (or alternatively `setPopupIcon`) allows refactoring the implementation of ribbon gallery expand button to display the expanded ribbon gallery in a popup without "faking" it as main action.

Instead of creating a popup menu panel with a customizer and then showing it via popup panel manager, the expand command is now configured with secondary content model, as well as a combination of various popup-related presentation model attributes.

Also fix the synchronization of selected command in expanded gallery and the in-ribbon gallery itself.

Fixes #452