More flexible logic for positioning the title pane text in ribbons that display contextual tabs
Consider ribbon task group headers as anchored title pane content that should be treated differently during the computation of the space available for the title.

If we have anchored content, compare horizontal space between leading and anchored, and anchored and trailing content. Whichever one is larger is "given" to display the window title.

Fixes #526